Migrates base_delivery_carrier_files to Odoo v8.0
Fields and function calls reimplemented according to new API.
Since stock module changed a little in 8.0 (stock.picking.out replaced
by stock.picking.type) - alter some logic.

Main changes include:

- checking against stock.picking.picking_type_id.code instead of
  stock.picking.out.type,
- add a file generation hook to _state_get function.

The former is needed because state field in the new api is of type
field.function which checks if all stock.move items are done. In such a
case stock.picking is moved to done too. Due to this reason, action_done
is not called most of the time and the file is not generated. File
generation is added just after the state checking function enabling the
desired functionality when action_done is not called.